+++ GradeTrackerPro/src/gradetrackerpro/graphics/screens/ScreenHome.java
@@ -47,8 +47,8 @@ public class ScreenHome extends JPanel implements ITrigger, IReceiver{
   }
   public void ping(String title, String[] data){
     if(title.equals("mouse-data")){
-      int x = Integer.parseInt(data[0]);
-      int y = Integer.parseInt(data[1]);
+      int x = (int)Double.parseDouble(data[0]);
+      int y = (int)Double.parseDouble(data[1]);
       int event = Integer.parseInt(data[2]);
       this.newCourseButton.mouseAction(x,y,event);
       this.loadCourseButton.mouseAction(x,y,event);
